Удален AsideMenuContainer
