+++ src/components/MainPage/MainPage.js
@@ -1,12 +1,12 @@
 import React from "react";
 import "./styles.scss";
-import AsideMenuContainer from "../common/AsideMenu/AsideMenuContainer";
+import AsideMenu from "../common/AsideMenu/AsideMenu";
 import Content from "./Content/Content";
 
 const MainPage = (props) => {
   return (
     <div className="main-page">
-      <AsideMenuContainer />
+      <AsideMenu />
       <Content />
     </div>
   );
